Move contexts to its own folder and perform refactoring.
