+++ src/Providers/contexts/ActiveSystemContext.js
@@ -1,6 +1,10 @@
 import {createContext} from "react";
 
 // This context stores the active system information.
-const ActiveSystemContext = createContext();
+const ActiveSystemContext = createContext({
+    systemId: null,
+    version: null,
+    deployment: null,
+});
 
 export default ActiveSystemContext;
